requirements.txt
appdirs==1.4.4
black==21.5b1
click==8.0.0
colorama==0.4.4
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
PySide2==5.15.2
PySimpleGUI==4.41.2
regex==2021.4.4
shiboken2==5.15.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
